fixed error: devide may return float but i was storing it in byte.

+++ homework 2/Task 1/Program.cs
@@ -38,7 +38,7 @@ namespace Task_1
             Console.WriteLine("-----------------------------------");
 
             // calculate average
-            byte average = (byte)((score1 + score2 + score3) / 3);
+            float average = score1 + score2 + score3 / 3;
 
             // log average
             Console.WriteLine("Average Score Of Students Is: " + average);
